unit test currency conversion and network connectivity

+++ app/src/main/java/com/example/currencyconverter/data/models/CurrencyResponse.kt
@@ -3,7 +3,7 @@ package com.example.currencyconverter.data.models
 data class CurrencyResponse(
     val base: String,
     val date: String,
-    val rates: Map<String, Float>,// change Rates to Map
+    val rates: Map<String, Double>,// change Rates to Map
     val success: Boolean,
     val timestamp: Int
 )
